+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'mymusiclist'
+rootProject.name = 'MyMusicList'
