+++ src/main/java/com/mymusiclist/backend/configuration/RedisConfiguration.java
@@ -3,6 +3,7 @@ package com.mymusiclist.backend.configuration;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
 import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
 import org.springframework.data.redis.core.RedisTemplate;
@@ -25,6 +26,7 @@ public class RedisConfiguration {
   }
 
   @Bean
+  @Primary
   public RedisTemplate<String, Object> redisTemplate() {
 
     RedisTemplate<String, Object> redisTemplate = new RedisTemplate<>();
