+++ src/main/java/com/mymusiclist/backend/post/service/CommentService.java
@@ -10,5 +10,4 @@ import org.springframework.stereotype.Service;
 public interface CommentService {
 
   List<CommentDto> getList(PostEntity postEntity);
-
 }
