+++ src/main/java/com/mymusiclist/backend/post/service/PostService.java
@@ -1,5 +1,6 @@
 package com.mymusiclist.backend.post.service;
 
+import com.mymusiclist.backend.post.dto.CommentDto;
 import com.mymusiclist.backend.post.dto.PostDetailDto;
 import com.mymusiclist.backend.post.dto.PostListDto;
 import com.mymusiclist.backend.post.dto.request.PostRequest;
