+++ src/test/java/com/mymusiclist/backend/MymusiclistApplicationTests.java
@@ -1,4 +1,4 @@
-package com.mymusiclist.mymusiclist;
+package com.mymusiclist.backend;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
